// src/app/models/transportista.ts
export class Transportista {
    idTransportista: string;
    nombres:         string;
    apellidos:       string;
    licencia:        string;
    placa:           string;
    vehiculo:        string;
    modelo:          string;
    idDistrito:      number;
    distrito:        string;
    seleccionado:   boolean = false;
}